refactor(gpr): move reduce_gpr -> reduce_model_gprs into the pipeline file
PR #69 review: reduce_gpr is pipeline-only (it needs essential_reacs + gene
KO/KI costs, which don't exist for a standalone compression), so relocate it
from networktools.py into compute_strain_designs.py and rename it
reduce_model_gprs for consistency with simplify_model_gprs. Drop the
remove_irrelevant_genes back-compat alias (no external callers). Its only
module-level dependency, evaluate_gpr_ast, stays in networktools (shared with
gene_kos_to_constraints) and is imported. Pure relocation + rename, design-
identical: test_04/05/07/08/10/11/12 (241 tests).
