+++ straindesign/networktools.py
@@ -499,7 +499,7 @@ def evaluate_gpr_ast(node, gene_states):
     """Evaluate a GPR AST node with given gene states.
 
     Supports arbitrary nesting of AND/OR operators (not limited to DNF/CNF).
-    Used by both gene_kos_to_constraints and reduce_gpr.
+    Used by both gene_kos_to_constraints and reduce_model_gprs.
 
     Args:
         node: An ast.Name or ast.BoolOp node from a parsed GPR rule
@@ -756,288 +756,6 @@ def resolve_gene_constraints(model, constraints):
         clean_constraints.extend(reaction_constraints)
 
     return clean_constraints
-
-
-def reduce_gpr(model, essential_reacs, gkis, gkos):
-    """Simplify GPR rules by removing non-targetable genes and reducing boolean expressions
-
-    This function is used in preprocessing of computational strain design computations. Often,
-    certain reactions, for instance, reactions essential for microbial growth can/must not be
-    targeted by interventions. That can be exploited to reduce the set of genes in which
-    interventions need to be considered.
-
-    Given a set of essential reactions that is to be maintained operational, some genes can be
-    removed from a metabolic model, either because they only affect only blocked reactions or
-    essential reactions, or because they are essential reactions and must not be removed. As a
-    consequence, the GPR rules of a model can be simplified using AST parsing for both DNF and non-DNF rules.
-
-
-    Example:
-        reduce_gpr(model, essential_reacs, gkis, gkos):
-    
-    Args:
-        model (cobra.Model):
-            A metabolic model that is an instance of the cobra.Model class containing GPR rules
-            
-        essential_reacs (list of str):
-            A list of identifiers of essential reactions.
-            
-        gkis, gkos (dict):
-            Dictionaries that contain the costs for gene knockouts and additions. E.g.,
-            gkos={'adhE': 1.0, 'ldhA' : 1.0 ...}
-            
-    Returns:
-        (dict):
-        An updated dictionary of the knockout costs in which irrelevant genes are removed.
-    """
-
-    def ast_to_gene_reaction_rule(node):
-        """
-        Convert an AST node back to gene reaction rule string format.
-        """
-        if isinstance(node, ast.Name):
-            return node.id
-        elif isinstance(node, ast.BoolOp):
-            child_strings = [ast_to_gene_reaction_rule(child) for child in node.values]
-            if isinstance(node.op, ast.And):
-                return ' and '.join(f'({s})' if ' or ' in s else s for s in child_strings)
-            elif isinstance(node.op, ast.Or):
-                return ' or '.join(f'({s})' if ' and ' in s else s for s in child_strings)
-        else:
-            raise ValueError(f"Unsupported AST node type: {type(node)}")
-
-    def simplify_gpr_ast(node, protected_genes_dict):
-        """
-        Simplify GPR AST by setting protected genes to True and applying boolean simplification.
-        This is equivalent to the original string-based approach but operates purely on AST.
-        """
-        return apply_gene_protection_to_ast(node, protected_genes_dict)
-
-    def apply_gene_protection_to_ast(node, protected_genes_dict):
-        """
-        Apply gene protection to AST by setting protected genes to True and simplifying boolean expressions.
-        Returns a simplified AST node with redundant terms removed and consistent gene ordering.
-        """
-        if isinstance(node, ast.Name):
-            if node.id in protected_genes_dict:
-                return True
-            else:
-                return node
-        elif isinstance(node, ast.BoolOp):
-            # Recursively apply to children
-            new_children = []
-            for child in node.values:
-                simplified_child = apply_gene_protection_to_ast(child, protected_genes_dict)
-
-                if isinstance(node.op, ast.And):
-                    if simplified_child is False:
-                        return False
-                    elif simplified_child is not True:
-                        new_children.append(simplified_child)
-                elif isinstance(node.op, ast.Or):
-                    if simplified_child is True:
-                        return True
-                    elif simplified_child is not False:
-                        new_children.append(simplified_child)
-
-            # Handle results
-            if not new_children:
-                return True if isinstance(node.op, ast.And) else False
-            elif len(new_children) == 1:
-                return new_children[0]
-            else:
-                # Apply additional simplifications for OR nodes
-                if isinstance(node.op, ast.Or):
-                    new_children = remove_redundant_or_terms(new_children)
-                    if len(new_children) == 1:
-                        return new_children[0]
-
-                # Sort children for consistent ordering (like string approach does)
-                sorted_children = sort_ast_nodes(new_children)
-                new_node = ast.BoolOp(op=node.op, values=sorted_children)
-                return new_node
-        else:
-            raise ValueError(f"Unsupported AST node type: {type(node)}")
-
-    def remove_redundant_or_terms(children):
-        """
-        Remove redundant terms from OR expressions using boolean logic simplification.
-        Example: (a and b and c) or (a and b) simplifies to (a and b)
-        since (a and b) is logically sufficient when both terms are present.
-        """
-        # Convert AST nodes to comparable forms
-        simplified = []
-        for child in children:
-            # Check if this child makes any other child redundant
-            is_redundant = False
-            for other in children:
-                if child is not other and is_subset_of(child, other):
-                    # child is a subset of other, so other is redundant
-                    is_redundant = False  # Keep child, remove other later
-                elif child is not other and is_subset_of(other, child):
-                    # other is a subset of child, so child is redundant
-                    is_redundant = True
-                    break
-            if not is_redundant:
-                simplified.append(child)
-
-        # Remove duplicates
-        unique = []
-        for child in simplified:
-            if not any(ast_nodes_equal(child, existing) for existing in unique):
-                unique.append(child)
-
-        return unique if unique else children
-
-    def is_subset_of(node1, node2):
-        """
-        Check if node1 logically absorbs node2 in boolean algebra.
-        
-        In OR expressions: A or (A and B) = A
-        This means A absorbs (A and B) because A is simpler/more general.
-        
-        For absorption to work: node1 must be "simpler" than node2,
-        meaning node2 implies node1 (node2 is more restrictive).
-        
-        Examples:
-        - mobA absorbs (mobA and mobB) 
-        - (a and b) absorbs (a and b and c)
-        """
-        # Case 1: Single gene absorbs AND expression containing that gene
-        if isinstance(node1, ast.Name) and isinstance(node2, ast.BoolOp) and isinstance(node2.op, ast.And):
-            genes_in_and = get_genes_from_ast(node2)
-            return node1.id in genes_in_and
-
-        # Case 2: Shorter AND expression absorbs longer AND expression with same genes
-        if (isinstance(node1, ast.BoolOp) and isinstance(node1.op, ast.And) and isinstance(node2, ast.BoolOp) and
-                isinstance(node2.op, ast.And)):
-            genes1 = get_genes_from_ast(node1)
-            genes2 = get_genes_from_ast(node2)
-            # node1 absorbs node2 if node1's genes are a proper subset of node2's genes
-            return genes1.issubset(genes2) and len(genes1) < len(genes2)
-
-        return False
-
-    def get_genes_from_ast(node):
-        """Extract set of genes from AST node"""
-        if isinstance(node, ast.Name):
-            return {node.id}
-        elif isinstance(node, ast.BoolOp):
-            genes = set()
-            for child in node.values:
-                genes.update(get_genes_from_ast(child))
-            return genes
-        return set()
-
-    def ast_nodes_equal(node1, node2):
-        """Check if two AST nodes are equivalent"""
-        if type(node1) != type(node2):
-            return False
-        if isinstance(node1, ast.Name):
-            return node1.id == node2.id
-        elif isinstance(node1, ast.BoolOp):
-            if type(node1.op) != type(node2.op):
-                return False
-            return (len(node1.values) == len(node2.values) and all(ast_nodes_equal(a, b) for a, b in zip(node1.values, node2.values)))
-        return False
-
-    def sort_ast_nodes(nodes):
-        """Sort AST nodes for consistent ordering"""
-
-        def node_sort_key(node):
-            if isinstance(node, ast.Name):
-                return (0, node.id)
-            elif isinstance(node, ast.BoolOp):
-                return (1, len(node.values), str(type(node.op)))
-            return (2, str(node))
-
-        return sorted(nodes, key=node_sort_key)
-
-    def is_gene_essential_to_reaction_ast(reaction, gene_id):
-        """
-        Determine if a gene is essential for a reaction using AST-based GPR analysis.
-        A gene is considered essential if removing it (setting it to False) makes 
-        the entire GPR expression evaluate to False, rendering the reaction impossible.
-        """
-        if not reaction.gene_reaction_rule:
-            return False
-
-        # Skip reactions without gene associations
-        if not reaction.gpr or not reaction.gpr.body:
-            return False
-
-        try:
-            # Test what happens if we knock out this gene using AST
-            gene_states = {gene_id: False}
-            result = evaluate_gpr_ast(reaction.gpr.body, gene_states)
-            return result is False
-        except Exception as e:
-            # Catch unsupported AST node types but don't fall back to string parsing
-            logging.warning(f'Unsupported AST node type in reaction {reaction.id} for gene {gene_id}: {e}')
-            return False
-
-    # 1) Remove gpr rules from blocked reactions
-    blocked_reactions = [reac.id for reac in model.reactions if reac.bounds == (0, 0)]
-    for rid in blocked_reactions:
-        model.reactions.get_by_id(rid).gene_reaction_rule = ''
-    for g in model.genes[::-1]:  # iterate in reverse order to avoid mixing up the order of the list when removing genes
-        if not g.reactions:
-            model.genes.remove(g)
-
-    protected_genes = set()
-
-    # 2. Protect genes that only occur in essential reactions
-    for g in model.genes:
-        if not g.reactions or {r.id for r in g.reactions}.issubset(essential_reacs):
-            protected_genes.add(g)
-
-    # 3. Protect genes that are essential to essential reactions (AST-based analysis)
-    for r in [model.reactions.get_by_id(s) for s in essential_reacs]:
-        for g in r.genes:
-            if is_gene_essential_to_reaction_ast(r, g.id):
-                protected_genes.add(g)
-
-    # 4. Remove essential genes, and knockouts without impact from gko_costs
-    [gkos.pop(pg.id) for pg in protected_genes if pg.id in gkos]
-
-    # 5. Add all not-knockable genes to the protected list
-    [protected_genes.add(g) for g in model.genes if (g.id not in gkos) and (g.name not in gkos)]  # support names or ids in gkos
-
-    # 6. genes with kiCosts are kept (remove from protected list so they can be targeted)
-    gki_ids = [g.id for g in model.genes if (g.id in gkis) or (g.name in gkis)]  # support names or ids in gkis
-    protected_genes = protected_genes.difference({model.genes.get_by_id(g) for g in gki_ids})
-    protected_genes_dict = {pg.id: True for pg in protected_genes}
-
-    # 7. Simplify GPR rules using AST-based boolean logic and remove non-targetable rules
-    for r in model.reactions:
-        if r.gene_reaction_rule and r.gpr and r.gpr.body:
-            try:
-                simplified = simplify_gpr_ast(r.gpr.body, protected_genes_dict)
-
-                if simplified is True:
-                    # Rule is always satisfied (cannot be knocked out)
-                    model.reactions.get_by_id(r.id).gene_reaction_rule = ''
-                elif simplified is False:
-                    # Rule is impossible - should not happen with proper protection
-                    logging.error(f'Something went wrong during gpr rule simplification for {r.id}.')
-                elif isinstance(simplified, (ast.Name, ast.BoolOp)):
-                    # Convert simplified AST back to string
-                    new_rule = ast_to_gene_reaction_rule(simplified)
-                    model.reactions.get_by_id(r.id).gene_reaction_rule = new_rule
-                # If simplified is the original node, keep original rule
-            except Exception as e:
-                logging.warning(f'Failed to simplify GPR rule for reaction {r.id}: {e}')
-
-    # 8. Remove obsolete genes and protected genes
-    for g in model.genes[::-1]:
-        if not g.reactions or g in protected_genes:
-            model.genes.remove(g)
-
-    return gkos
-
-
-# backward-compat alias
-remove_irrelevant_genes = reduce_gpr
 
 
 def extend_model_gpr(model, use_names=False):
